requirements.txt
flask==0.11.1
flask-sqlalchemy==2.1
sqlalchemy-migrate==0.10.0
pymysql==0.7.9
marshmallow==2.10.3
flask-marshmallow==0.7.0
marshmallow-sqlalchemy==0.12.0
